requirements.txt
streamlit
streamlit-agraph
pandas
openai
python-dotenv
